backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
cloudinary
python-dotenv
ultralytics
# for web backend (signup/login)
pymongo
bcrypt
python-jose[cryptography]
